@@ -7,20 +7,18 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "history": "^5.2.0",
     "jsurl": "0.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6.3.0",
-    "react-router": "6.3.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^2.2.1",
-    "@types/node": "14.x",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@vitejs/plugin-react": "^1.0.1",
-    "typescript": "^4.3.5",
-    "vite": "^2.5.0"
+    "@rollup/plugin-replace": "^5.0.2",
+    "@types/node": "18.x",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react": "^3.0.1",
+    "typescript": "^4.9.5",
+    "vite": "^4.0.4"
   }
 }
