@@ -12,20 +12,18 @@
   "dependencies": {
     "compression": "1.7.4",
     "cross-env": "^7.0.3",
-    "express": "^4.17.1",
-    "history": "^5.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6.3.0",
-    "react-router": "6.3.0"
+    "express": "^4.18.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
-    "@types/node": "^16.10.2",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@vitejs/plugin-react": "^1.0.1",
-    "typescript": "^4.3.5",
-    "vite": "^2.5.0"
+    "@rollup/plugin-replace": "^5.0.2",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react": "^3.0.1",
+    "typescript": "^4.9.5",
+    "vite": "^4.0.4"
   }
 }
