@@ -14,7 +14,8 @@ async function createServer() {
   if (!isProduction) {
     vite = await require("vite").createServer({
       root: process.cwd(),
-      server: { middlewareMode: "ssr" },
+      server: { middlewareMode: true },
+      appType: "custom",
     });
 
     app.use(vite.middlewares);
