@@ -5,6 +5,9 @@ import react from "@vitejs/plugin-react";
 import rollupReplace from "@rollup/plugin-replace";
 
 export default defineConfig({
+  server: {
+    port: 3000,
+  },
   plugins: [
     rollupReplace({
       preventAssignment: true,
@@ -27,9 +30,13 @@ export default defineConfig({
   resolve: process.env.USE_SOURCE
     ? {
         alias: {
+          "@remix-run/router": path.resolve(
+            __dirname,
+            "../../packages/router/index.ts"
+          ),
           "react-router": path.resolve(
             __dirname,
-            "../../packages/react-router/index.tsx"
+            "../../packages/react-router/index.ts"
           ),
           "react-router-dom": path.resolve(
             __dirname,
