@@ -25,7 +25,7 @@
  * - TSC needs to generate the types, and it has to derive the output paths from
  *   the import paths. If we have a weird require *outside of this package* to
  *   "../../react-router-dom" it's going to generate types from the common root
- *   of all module paths (Which makes sense becuase what else would it do? It
+ *   of all module paths (Which makes sense because what else would it do? It
  *   needs to write the type files next to the source files so that typescript
  *   can resolve the types for tooling in the same location as the modules).
  *   Because tsc isn't as flexible as rollup, we have no control over this
@@ -47,31 +47,74 @@
  *   deprecate the deep require if we wanted to avoid the duplication here.
  */
 export type {
+  ActionFunction,
+  ActionFunctionArgs,
+  AwaitProps,
+  BrowserRouterProps,
+  DataRouteMatch,
+  DataRouteObject,
+  ErrorResponse,
+  Fetcher,
+  FetcherWithComponents,
+  FormEncType,
+  FormMethod,
+  FormProps,
+  FutureConfig,
+  GetScrollRestorationKeyFunction,
   Hash,
+  HashRouterProps,
+  HistoryRouterProps,
+  IndexRouteObject,
+  IndexRouteProps,
+  JsonFunction,
+  LayoutRouteProps,
+  LinkProps,
+  LoaderFunction,
+  LoaderFunctionArgs,
   Location,
-  Path,
-  To,
   MemoryRouterProps,
+  NavLinkProps,
   NavigateFunction,
   NavigateOptions,
   NavigateProps,
+  Navigation,
   Navigator,
+  NonIndexRouteObject,
   OutletProps,
+  ParamKeyValuePair,
+  ParamParseKey,
   Params,
+  Path,
   PathMatch,
+  PathParam,
+  PathPattern,
+  PathRouteProps,
+  Pathname,
+  RedirectFunction,
+  RelativeRoutingType,
   RouteMatch,
   RouteObject,
   RouteProps,
-  PathRouteProps,
-  LayoutRouteProps,
-  IndexRouteProps,
   RouterProps,
-  Pathname,
-  Search,
+  RouterProviderProps,
   RoutesProps,
+  ScrollRestorationProps,
+  Search,
+  ShouldRevalidateFunction,
+  ShouldRevalidateFunctionArgs,
+  SubmitFunction,
+  SubmitOptions,
+  To,
+  URLSearchParamsInit,
+  UIMatch,
+  Blocker,
+  BlockerFunction,
 } from "./react-router-dom";
 export {
+  AbortedDeferredError,
+  Await,
   BrowserRouter,
+  Form,
   HashRouter,
   Link,
   MemoryRouter,
@@ -81,34 +124,65 @@ export {
   Outlet,
   Route,
   Router,
+  RouterProvider,
   Routes,
+  ScrollRestoration,
+  UNSAFE_DataRouterContext,
+  UNSAFE_DataRouterStateContext,
   UNSAFE_LocationContext,
   UNSAFE_NavigationContext,
   UNSAFE_RouteContext,
+  UNSAFE_useRouteId,
+  UNSAFE_useScrollRestoration,
+  createBrowserRouter,
+  createHashRouter,
+  createMemoryRouter,
   createPath,
   createRoutesFromChildren,
+  createRoutesFromElements,
   createSearchParams,
+  defer,
   generatePath,
+  isRouteErrorResponse,
+  json,
   matchPath,
   matchRoutes,
   parsePath,
+  redirect,
+  redirectDocument,
   renderMatches,
   resolvePath,
   unstable_HistoryRouter,
+  useBlocker,
+  unstable_usePrompt,
+  useActionData,
+  useAsyncError,
+  useAsyncValue,
+  useBeforeUnload,
+  useFetcher,
+  useFetchers,
+  useFormAction,
   useHref,
   useInRouterContext,
   useLinkClickHandler,
+  useLoaderData,
   useLocation,
   useMatch,
+  useMatches,
   useNavigate,
+  useNavigation,
   useNavigationType,
   useOutlet,
   useOutletContext,
   useParams,
   useResolvedPath,
+  useRevalidator,
+  useRouteError,
+  useRouteLoaderData,
   useRoutes,
   useSearchParams,
+  useSubmit,
 } from "./react-router-dom";
 
 export type { StaticRouterProps } from "./lib/components";
-export { CompatRouter, CompatRoute, StaticRouter } from "./lib/components";
+export { CompatRoute, CompatRouter, StaticRouter } from "./lib/components";
