@@ -14,14 +14,25 @@ import { Router, Routes, Route } from "../react-router-dom";
 // v5 isn't in TypeScript, they'll also lose the @types/react-router with this
 // but not worried about that for now.
 export function CompatRoute(props: any) {
-  let { path } = props;
+  let { location, path } = props;
   if (!props.exact) path += "/*";
   return (
-    <Routes>
+    <Routes location={location}>
       <Route path={path} element={<RouteV5 {...props} />} />
     </Routes>
   );
 }
+
+// Copied with 💜 from https://github.com/bvaughn/react-resizable-panels/blob/main/packages/react-resizable-panels/src/hooks/useIsomorphicEffect.ts
+const canUseEffectHooks = !!(
+  typeof window !== "undefined" &&
+  typeof window.document !== "undefined" &&
+  typeof window.document.createElement !== "undefined"
+);
+
+const useIsomorphicLayoutEffect = canUseEffectHooks
+  ? React.useLayoutEffect
+  : () => {};
 
 export function CompatRouter({ children }: { children: React.ReactNode }) {
   let history = useHistory();
@@ -30,7 +41,7 @@ export function CompatRouter({ children }: { children: React.ReactNode }) {
     action: history.action,
   }));
 
-  React.useLayoutEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     history.listen((location: Location, action: Action) =>
       setState({ location, action })
     );
@@ -55,8 +66,10 @@ export interface StaticRouterProps {
   location: Partial<Location> | string;
 }
 
+const ABSOLUTE_URL_REGEX = /^(?:[a-z][a-z0-9+.-]*:|\/\/)/i;
+
 /**
- * A <Router> that may not transition to any other location. This is useful
+ * A `<Router>` that may not navigate to any other location. This is useful
  * on the server where there is no stateful UI.
  */
 export function StaticRouter({
@@ -80,6 +93,21 @@ export function StaticRouter({
   let staticNavigator = {
     createHref(to: To) {
       return typeof to === "string" ? to : createPath(to);
+    },
+    encodeLocation(to: To) {
+      let href = typeof to === "string" ? to : createPath(to);
+      // Treating this as a full URL will strip any trailing spaces so we need to
+      // pre-encode them since they might be part of a matching splat param from
+      // an ancestor route
+      href = href.replace(/ $/, "%20");
+      let encoded = ABSOLUTE_URL_REGEX.test(href)
+        ? new URL(href)
+        : new URL(href, "http://localhost");
+      return {
+        pathname: encoded.pathname,
+        search: encoded.search,
+        hash: encoded.hash,
+      };
     },
     push(to: To) {
       throw new Error(
