@@ -20,11 +20,11 @@ describe("A <StaticRouter>", () => {
         </StaticRouter>
       );
 
-      expect(location).toMatchObject({
+      expect(location).toEqual({
         pathname: "/the/path",
         search: "?the=query",
         hash: "#the-hash",
-        state: {},
+        state: null,
         key: expect.any(String),
       });
     });
@@ -48,11 +48,11 @@ describe("A <StaticRouter>", () => {
         </StaticRouter>
       );
 
-      expect(location).toMatchObject({
+      expect(location).toEqual({
         pathname: "/the/path",
         search: "?the=query",
         hash: "",
-        state: {},
+        state: null,
         key: expect.any(String),
       });
     });
