@@ -1,8 +1,9 @@
+/** @type {import('jest').Config} */
 module.exports = {
   preset: "react-native",
   testMatch: ["**/__tests__/*-test.[jt]s?(x)"],
   transform: {
-    "\\.[jt]sx?$": "<rootDir>/node_modules/react-native/jest/preprocessor.js",
+    "\\.[jt]sx?$": "./jest-transformer.js",
   },
   globals: {
     __DEV__: true,
@@ -11,4 +12,16 @@ module.exports = {
     "<rootDir>/node_modules", // for react-native
   ],
   setupFiles: ["<rootDir>/__tests__/setup.ts"],
+  moduleNameMapper: {
+    "^@remix-run/router$": "<rootDir>/../router/index.ts",
+    "^@remix-run/web-blob$": require.resolve("@remix-run/web-blob"),
+    "^@remix-run/web-fetch$": require.resolve("@remix-run/web-fetch"),
+    "^@remix-run/web-form-data$": require.resolve("@remix-run/web-form-data"),
+    "^@remix-run/web-stream$": require.resolve("@remix-run/web-stream"),
+    "^@web3-storage/multipart-parser$": require.resolve(
+      "@web3-storage/multipart-parser"
+    ),
+    "^react-router$": "<rootDir>/../react-router/index.ts",
+    "^react-router-native$": "<rootDir>/index.tsx",
+  },
 };
