@@ -8,9 +8,10 @@ describe("index route matching", () => {
           {
             path: "/users",
             children: [
+              // This config is not valid because index routes cannot have children
+              // @ts-expect-error
               {
                 index: true,
-                // This config is not valid because index routes cannot have children
                 children: [{ path: "not-valid" }],
               },
               { path: ":id" },
@@ -19,6 +20,8 @@ describe("index route matching", () => {
         ],
         "/users/mj"
       );
-    }).toThrow("must not have child routes");
+    }).toThrowErrorMatchingInlineSnapshot(
+      `"Index routes must not have child routes. Please remove all child routes from route path "/users/"."`
+    );
   });
 });
