@@ -1,6 +1,7 @@
 import * as React from "react";
 import * as TestRenderer from "react-test-renderer";
-import { MemoryRouter, PathMatch, Routes, Route, useMatch } from "react-router";
+import type { PathMatch } from "react-router";
+import { MemoryRouter, Routes, Route, useMatch } from "react-router";
 
 function ShowMatch({ pattern }: { pattern: string }) {
   return <pre>{JSON.stringify(useMatch(pattern), null, 2)}</pre>;
