@@ -1,40 +1,19 @@
+// FIXME: @remix-run/router isn't being automatically versioned
 const path = require("path");
 const { execSync } = require("child_process");
 const fsp = require("fs/promises");
 const chalk = require("chalk");
-const Confirm = require("prompt-confirm");
-const jsonfile = require("jsonfile");
 const semver = require("semver");
 
-const rootDir = path.resolve(__dirname, "..");
-const examplesDir = path.resolve(rootDir, "examples");
-
-/**
- * @param {string} packageName
- * @param {string} [directory]
- * @returns {string}
- */
-function packageJson(packageName, directory = "packages") {
-  return path.join(rootDir, directory, packageName, "package.json");
-}
-
-/**
- * @param {*} cond
- * @param {string} message
- * @returns {asserts cond}
- */
-function invariant(cond, message) {
-  if (!cond) throw new Error(message);
-}
-
-function ensureCleanWorkingDirectory() {
-  let status = execSync(`git status --porcelain`).toString().trim();
-  let lines = status.split("\n");
-  invariant(
-    lines.every((line) => line === "" || line.startsWith("?")),
-    "Working directory is not clean. Please commit or stash your changes."
-  );
-}
+const {
+  ensureCleanWorkingDirectory,
+  getPackageVersion,
+  invariant,
+  prompt,
+  updateExamplesPackageConfig,
+  updatePackageConfig,
+} = require("./utils");
+const { EXAMPLES_DIR } = require("./constants");
 
 /**
  * @param {string} currentVersion
@@ -55,52 +34,18 @@ function getNextVersion(currentVersion, givenVersion, prereleaseId) {
     );
   }
 
-  let nextVersion = semver.inc(currentVersion, givenVersion, prereleaseId);
+  let nextVersion;
+  if (givenVersion === "experimental") {
+    let hash = execSync(`git rev-parse --short HEAD`).toString().trim();
+    nextVersion = `0.0.0-experimental-${hash}`;
+  } else {
+    // @ts-ignore
+    nextVersion = semver.inc(currentVersion, givenVersion, prereleaseId);
+  }
 
   invariant(nextVersion != null, `Invalid version specifier: ${givenVersion}`);
 
   return nextVersion;
-}
-
-/**
- * @param {string} question
- * @returns {Promise<string>}
- */
-async function prompt(question) {
-  let confirm = new Confirm(question);
-  let answer = await confirm.run();
-  return answer;
-}
-
-/**
- * @param {string} packageName
- */
-async function getPackageVersion(packageName) {
-  let file = packageJson(packageName);
-  let json = await jsonfile.readFile(file);
-  return json.version;
-}
-
-/**
- * @param {string} packageName
- * @param {(json: string) => any} transform
- */
-async function updatePackageConfig(packageName, transform) {
-  let file = packageJson(packageName);
-  let json = await jsonfile.readFile(file);
-  transform(json);
-  await jsonfile.writeFile(file, json, { spaces: 2 });
-}
-
-/**
- * @param {string} example
- * @param {(json: string) => any} transform
- */
-async function updateExamplesPackageConfig(example, transform) {
-  let file = packageJson(example, "examples");
-  let json = await jsonfile.readFile(file);
-  transform(json);
-  await jsonfile.writeFile(file, json, { spaces: 2 });
 }
 
 async function run() {
@@ -108,16 +53,21 @@ async function run() {
     let args = process.argv.slice(2);
     let givenVersion = args[0];
     let prereleaseId = args[1];
+    let isExperimental = givenVersion === "experimental";
 
     // 0. Make sure the working directory is clean
     ensureCleanWorkingDirectory();
 
     // 1. Get the next version number
+    let currentRouterVersion = await getPackageVersion("router");
     let currentVersion = await getPackageVersion("react-router");
     let version = semver.valid(givenVersion);
     if (version == null) {
       version = getNextVersion(currentVersion, givenVersion, prereleaseId);
     }
+
+    // We will only bump the router version if this is an experimental
+    let routerVersion = currentRouterVersion;
 
     // 2. Confirm the next version number
     let answer = await prompt(
@@ -126,15 +76,34 @@ async function run() {
 
     if (answer === false) return 0;
 
+    // We only handle @remix-run/router for experimental since in normal/pre
+    // releases it's versioned independently from the rest of the packages
+    if (isExperimental) {
+      routerVersion = version;
+      // 2.5. Update @remix-run/router version
+      await updatePackageConfig("router", (config) => {
+        config.version = routerVersion;
+      });
+      console.log(
+        chalk.green(`  Updated @remix-run/router to version ${version}`)
+      );
+    }
+
     // 3. Update react-router version
     await updatePackageConfig("react-router", (config) => {
       config.version = version;
+      if (isExperimental) {
+        config.dependencies["@remix-run/router"] = routerVersion;
+      }
     });
     console.log(chalk.green(`  Updated react-router to version ${version}`));
 
     // 4. Update react-router-dom version + react-router dep
     await updatePackageConfig("react-router-dom", (config) => {
       config.version = version;
+      if (isExperimental) {
+        config.dependencies["@remix-run/router"] = routerVersion;
+      }
       config.dependencies["react-router"] = version;
     });
     console.log(
@@ -160,14 +129,21 @@ async function run() {
     );
 
     // 6. Update react-router and react-router-dom versions in the examples
-    let examples = await fsp.readdir(examplesDir);
+    let examples = await fsp.readdir(EXAMPLES_DIR);
     for (const example of examples) {
-      let stat = await fsp.stat(path.join(examplesDir, example));
+      let stat = await fsp.stat(path.join(EXAMPLES_DIR, example));
       if (!stat.isDirectory()) continue;
 
       await updateExamplesPackageConfig(example, (config) => {
-        config.dependencies["react-router"] = version;
-        config.dependencies["react-router-dom"] = version;
+        if (config.dependencies["@remix-run/router"]) {
+          config.dependencies["@remix-run/router"] = routerVersion;
+        }
+        if (config.dependencies["react-router"]) {
+          config.dependencies["react-router"] = version;
+        }
+        if (config.dependencies["react-router-dom"]) {
+          config.dependencies["react-router-dom"] = version;
+        }
       });
     }
 
@@ -175,6 +151,14 @@ async function run() {
     execSync(`git commit --all --message="Version ${version}"`);
     execSync(`git tag -a -m "Version ${version}" v${version}`);
     console.log(chalk.green(`  Committed and tagged version ${version}`));
+
+    if (givenVersion !== "experimental") {
+      console.log(
+        chalk.red(
+          `  🚨 @remix-run/router isn't handled by this script, do it manually!`
+        )
+      );
+    }
   } catch (error) {
     console.log();
     console.error(chalk.red(`  ${error.message}`));
